fix(chord/delta): build on 32-bit targets (verification of upstream 9a139c62b port)
`GOOS=linux GOARCH=386 go build ./...` failed in chord/delta alone:
arrayIndexOf and canonicalIndex compared an int with the untyped
constant maxArrayIndex (2^32-2), which overflows a 32-bit int. The
comparisons now widen the non-negative operand to uint64, so the
constant never has to fit in an int. Behaviour on 64-bit is unchanged.
The old track.go:1341 had the same defect before the port.

Red: the 386 build failed at draft.go:494 and value.go:263 ("maxArrayIndex
(untyped int constant 4294967294) overflows int"). It now builds, and the
module's non-test build is clean on 386. Test files across the module
(client, server, coding, chord) use 64-bit constants, as before.

+++ chord/delta/draft.go
@@ -491,7 +491,7 @@ func (d *Draft) At(key any) *Draft {
 func arrayIndexOf(seg Seg) (int, bool) {
 	switch s := seg.(type) {
 	case Index:
-		if s >= 0 && s <= maxArrayIndex {
+		if s >= 0 && uint64(s) <= maxArrayIndex {
 			return int(s), true
 		}
 	case Key:
